% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses.R
\name{dualFunctionB_q_1}
\alias{dualFunctionB_q_1}
\title{dualFunctionB_q_1}
\usage{
dualFunctionB_q_1(coeff1, coeff2, x)
}
\arguments{
\item{coeff1}{coefficients (A,B,C,D,E,F) in objective function Ax^2 + Bxy + Cy^2 + Dx + Ey + F}

\item{coeff2}{coefficients (A,B,C,D,E,F) in constraint Ax^2 + Bxy + Cy^2 + Dx + Ey + F = 0}

\item{x}{vector x for plot}
}
\value{
the dual function to problem B with one constraint
}
\description{
Building the dual function in problem B with one constraint
}
\examples{
dualFunctionB_q_1(c(10,-5,2,4,2,-6), c(10,-5,2,4,2,-6), seq(-5,5,0.01))
}
