% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{globalCost}
\alias{globalCost}
\title{globalCost}
\usage{
globalCost(data, chpts, beta)
}
\arguments{
\item{data}{a dataframe with two components: y1 and y2, time series of same length}

\item{chpts}{vector of change-points obtained by one of the OP algorithl using data-points 'data'}

\item{beta}{penalty}
}
\value{
the global cost value (minimal value for the optimization problem)
}
\description{
Computing the global cost
}
\examples{
data <- dataGenerator2D(chpts = c(30,100,120), means1 = c(0,1,0), means2 = c(7,1,-4))
res <- OP_2D(data)
globalCost(data, res$changepoints, 4*log(nrow(data)))
}
