% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_cost_functions.R
\name{globalCost_Reg}
\alias{globalCost_Reg}
\title{globalCost_Reg}
\usage{
globalCost_Reg(data, chpts, beta)
}
\arguments{
\item{data}{a dataframe with two components: x and y, time series of same length}

\item{chpts}{vector of change-points obtained by one of the OP algorithm (Reg case) using data-points 'data'}

\item{beta}{penalty value}
}
\value{
the global cost value (minimal value for the optimization problem)
}
\description{
Computing the global cost of the segmentation for linear regression problem
}
\examples{
data <- dataGenerator_Reg(chpts = c(40,90), A = c(2,-1),  B = c(-1,2), meansX = c(1,2))
res <- OP_Reg(data)
globalCost_Reg(data, res$changepoints, 4*log(nrow(data)))
}
