% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{globalCost_Slope}
\alias{globalCost_Slope}
\title{globalCost_Slope}
\usage{
globalCost_Slope(data, chpts, kinks, beta)
}
\arguments{
\item{data}{a dataframe with two components: y1 and y2, time series of same length}

\item{chpts}{vector of change-points obtained by one of the OP algorithm (2D case) using data-points 'data'}

\item{kinks}{vector of successive kink values (kink heights)}

\item{beta}{penalty value}
}
\value{
the global cost value (minimal value for the optimization problem)
}
\description{
Computing the global cost of the segmentation for slope problem
}
