% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses.R
\name{my.coeff.ellipse.from.data}
\alias{my.coeff.ellipse.from.data}
\title{my.coeff.ellipse.from.data}
\usage{
my.coeff.ellipse.from.data(x, y, level)
}
\arguments{
\item{x}{data vector dim 1}

\item{y}{data vector dim 2 (linked to x by simple linear regression)}

\item{level}{a value equal to m_{k_1} - m_{j_1}}
}
\value{
A list of elements
}
\description{
Building coefficients for inscribed and circumscribed circles to an ellipse from real vector data (x,y)
}
\examples{
new.coeff.for.circle(c(10,-5,2,4,2,-6), "in")
}
