% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses.R
\name{my.ellipse}
\alias{my.ellipse}
\title{my.ellipse}
\usage{
my.ellipse(coeff, n.points = 1000)
}
\arguments{
\item{coeff}{coefficients (A,B,C,D,E,F) in equation Ax^2 + Bxy + Cy^2 + Dx + Ey + F = 0}

\item{n.points}{number of (x,y) points to generate}
}
\value{
matrix of size "n.points x 2" for (x,y) bipoints on the 2d plane describing an ellipse
}
\description{
Generating 2d points describing an ellipse from its coefficients in Cartesian form.
}
\examples{
my.ellipse(c(10,-5,2,4,2,-6), 1000)
}
