% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses.R
\name{new.coeff.for.circle}
\alias{new.coeff.for.circle}
\title{new.coeff.for.circle}
\usage{
new.coeff.for.circle(coeff, type = "in")
}
\arguments{
\item{coeff}{coefficients (A,B,C,D,E,F) in equation Ax^2 + Bxy + Cy^2 + Dx + Ey + F = 0}

\item{type}{"in" or "out"}
}
\value{
Six coefficients for Ax^2 + Bxy + Cy^2 + Dx + Ey + F = 0 (A=C and B=0)
}
\description{
Building coefficients for inscribed and circumscribed circles to an ellipse whose coefficients are given in Cartesian form.
}
\examples{
new.coeff.for.circle(c(10,-5,2,4,2,-6), "in")
}
