% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_bivariate_independent.R
\name{OP_2D}
\alias{OP_2D}
\title{OP_2D}
\usage{
OP_2D(data, beta = 4 * log(nrow(data)))
}
\arguments{
\item{data}{a dataframe with two components: y1 and y2, time series of same length}

\item{beta}{penalty value}
}
\value{
a list with the change-point elements (each last index of each segment)
}
\description{
Optimal Partitioning algorithm for bivariate independent time series (no pruning)
}
\examples{
OP_2D(dataGenerator_2D(chpts = c(30,100,120), means1 = c(0,5,0), means2 = c(7,1,-4)))
}
