% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_bivariate_independent.R
\name{OP_2D_1C}
\alias{OP_2D_1C}
\title{OP_2D_1C}
\usage{
OP_2D_1C(data, beta = 4 * log(length(data$y1)))
}
\arguments{
\item{data}{a dataframe with two components: y1 and y2, time series of same length}

\item{beta}{penalty}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration
}
\description{
Optimal Partitioning algorithm for bivariate independent time series (with OP1C algorithm)
}
\examples{
OP_2D_1C(dataGenerator2D(chpts = c(30,100,120), means1 = c(0,1,0), means2 = c(7,1,-4)))
}
