% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_2D_2C.R, R/OP_2D_2C2.R
\name{OP_2D_2C}
\alias{OP_2D_2C}
\title{OP_2D_2C}
\usage{
OP_2D_2C(data, beta = 4 * log(nrow(data)), testMode = 2)

OP_2D_2C(data, beta = 4 * log(nrow(data)), testMode = 2)
}
\arguments{
\item{data}{a dataframe with two components: y1 and y2, time series of same length}

\item{beta}{penalty value}

\item{testMode}{inner pruning test modes (0, 1 or 2)}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration

a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration
}
\description{
Optimal Partitioning algorithm for bivariate independent time series (with OP2C algorithm)

Optimal Partitioning algorithm for bivariate independent time series (with OP2C algorithm)
}
\examples{
OP_2D_2C(dataGenerator_2D(chpts = c(3,10,12), means1 = c(0,1,0), means2 = c(7,1,-4)))
OP_2D_2C(dataGenerator_2D(chpts = c(3,10,12), means1 = c(0,1,0), means2 = c(7,1,-4)))
}
