% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_2D_AC.R
\name{OP_2D_AC}
\alias{OP_2D_AC}
\title{OP_2D_AC}
\usage{
OP_2D_AC(data, beta = 4 * log(nrow(data)))
}
\arguments{
\item{data}{a data-frame with two components: y1 and y2, time series of same length}

\item{beta}{penalty value}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration
}
\description{
Optimal Partitioning algorithm for bivariate independent time series (with OPAC algorithm)
}
