% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_bivariate_independent_fast.R
\name{OP_2D_PELT_fast}
\alias{OP_2D_PELT_fast}
\title{OP_2D_PELT_fast}
\usage{
OP_2D_PELT_fast(data, beta = 4 * log(nrow(data)))
}
\arguments{
\item{data}{a dataframe with two components: y1 and y2, time series of same length}

\item{beta}{penalty value}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration
}
\description{
Optimal Partitioning algorithm for bivariate independent time series (with PELT pruning)
}
\examples{
OP_2D_PELT_fast(dataGenerator2D(chpts = c(30,100,120), means1 = c(0,1,0), means2 = c(7,1,-4)))
}
