% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_Reg_noPruningAndPELT.R
\name{OP_Reg}
\alias{OP_Reg}
\title{OP_Reg}
\usage{
OP_Reg(data, beta = 4 * log(nrow(data)))
}
\arguments{
\item{data}{a data-frame with two components: x and y, time series of same length}

\item{beta}{penalty value}
}
\value{
a list with the change-point elements (each last index of each segment)
}
\description{
Optimal Partitioning algorithm for changes in simple regression (no pruning)
}
\examples{
OP_Reg(dataGenerator_Reg(chpts = c(50,100,150), A = c(-1,1,-1), B = c(-1,1,-1)))
}
