% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_simpleRegression.R
\name{OP_Reg_1C}
\alias{OP_Reg_1C}
\title{OP_Reg_1C}
\usage{
OP_Reg_1C(data, beta = 4 * log(nrow(data)))
}
\arguments{
\item{data}{a dataframe with two components: x and y, time series of same length}

\item{beta}{penalty value}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration
}
\description{
Optimal Partitioning algorithm for changes in simple regression (with OP1C pruning)
}
\examples{
OP_Reg_1C(dataGeneratorRegression(chpts = c(50,100,150), A = c(-1,1,-1), B = c(-1,1,-1)))
}
