% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_Reg_1C_Approx.R
\name{OP_Reg_1C_Approx}
\alias{OP_Reg_1C_Approx}
\title{OP_Reg_1C_Approx}
\usage{
OP_Reg_1C_Approx(data, beta = 4 * log(nrow(data)))
}
\arguments{
\item{data}{a data-frame with two components: x and y, time series of same length}

\item{beta}{penalty value}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration
}
\description{
Optimal Partitioning algorithm for change in simple regression (with OP1C algorithm and circle approximation)
}
