% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipses.R
\name{points.on.ellipse}
\alias{points.on.ellipse}
\title{points.on.ellipse}
\usage{
\method{points}{on.ellipse}(coeff, n.points = 1000)
}
\arguments{
\item{coeff}{coefficients (A,B,C,D,E,F) in equation Ax^2 + 2Bxy + Cy^2 + 2Dx + 2Ey + F = 0}

\item{n.points}{number of (x,y) points to generate}
}
\value{
matrix of size "2*n.points x 2" for (x,y) bipoints on the 2d plane describing an ellipse
}
\description{
Generating 2d points describing an ellipse from its coefficients in Cartesian form.
}
