% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{dataGenerator2D}
\alias{dataGenerator2D}
\title{dataGenerator2D}
\usage{
dataGenerator2D(
  chpts = 100,
  means1 = 0,
  means2 = 0,
  sdNoise1 = 1,
  sdNoise2 = 1,
  type = "gauss"
)
}
\arguments{
\item{chpts}{a vector of increasing change-point indices}

\item{means1}{vector of successive segment means for time series 1}

\item{means2}{vector of successive segment means for time series 2}

\item{sdNoise1}{standard deviation of the Gaussian noise for time series 1}

\item{sdNoise2}{standard deviation of the Gaussian noise for time series 2}

\item{type}{the model: gauss or poisson}
}
\value{
a dataframe with two simulated time series with same length in component y1 and y2
}
\description{
Generating data for change-point detection in bivariate independent time series
}
\examples{
myData <- dataGenerator2D(chpts = c(30,100,120), means1 = c(0,5,0), means2 = c(7,1,-4))
}
