% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{dataGeneratorRegression}
\alias{dataGeneratorRegression}
\title{dataGeneratorRegression}
\usage{
dataGeneratorRegression(
  chpts = c(50, 150),
  A = c(2, 1),
  B = c(-1, 2),
  meansX = 0,
  sdX = 1,
  sdNoise = 1
)
}
\arguments{
\item{chpts}{a vector of increasing change-point indices}

\item{A}{vector of regression coefficients A in Ax+B simple regression model}

\item{B}{vector of regression coefficients B in Ax+B simple regression model}

\item{meansX}{mean values for X values generated by a Gaussian model}

\item{sdX}{standard deviation values for X values generated by a Gaussian model}

\item{sdNoise}{standard deviation of the Gaussian noise}
}
\value{
a dataframe with time series x and y for change-points in regression of type y = Ax + B + noise
}
\description{
Generate data for changes in bivariate independent time series
}
\examples{
myData <- dataGeneratorRegression(chpts = c(40,90))
}
