% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{dataGeneratorSlope}
\alias{dataGeneratorSlope}
\title{dataGeneratorSlope}
\usage{
dataGeneratorSlope(chpts, kinks, varNoise = 0)
}
\arguments{
\item{chpts}{a vector of increasing changepoint indices}

\item{kinks}{vector of successive kink values (kink height)}

\item{varNoise}{variance of the Gaussian noise}
}
\value{
a vector of (univariate) simulated data following the (chpts, kinks) 2D data-points.
}
\description{
Generate data with a given continuous piecewise linear model
}
\examples{
myData <- dataGeneratorSlope(c(1,30,100,150), c(-2,pi,0.001,2.1), 0.1)
}
