% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_2D_AC.R
\name{OP_2D_AC}
\alias{OP_2D_AC}
\title{OP_2D_2C}
\usage{
OP_2D_AC(data, beta = 4 * log(nrow(data)), testMode = 2)
}
\arguments{
\item{data}{a dataframe with two components: y1 and y2, time series of same length}

\item{beta}{penalty value}

\item{testMode}{inner pruning test modes (0, 1 or 2)}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration
}
\description{
Optimal Partitioning algorithm for bivariate independent time series (with OP2C algorithm)
}
