% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_Reg_1C.R
\name{OP_Reg_1C}
\alias{OP_Reg_1C}
\title{OP_Reg_1C}
\usage{
OP_Reg_1C(data, beta = 4 * log(nrow(data)))
}
\arguments{
\item{data}{a data-frame with two components: y1 and y2, time series of same length}

\item{beta}{penalty value}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector nb counting the number of non-pruned elements at each iteration
}
\description{
Optimal Partitioning algorithm for bivariate independent time series (with OP1C algorithm)
}
